+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/next/FileSystemNextExceptions.kt
@@ -29,6 +29,16 @@ internal class InvalidPermissionException(permission: Permission) :
     "Missing '${permission.name}' permission for accessing the file."
   )
 
+internal class UnableToReadHandleException(reason: String) :
+  CodedException(
+    "Unable to read from a file handle: '$reason'"
+  )
+
+internal class UnableToWriteHandleException(reason: String) :
+  CodedException(
+    "Unable to write to a file handle: '$reason'"
+  )
+
 internal class DestinationAlreadyExistsException :
   CodedException(
     "Destination already exists"