+++ packages/expo-file-system/ios/Next/FileSystemNextExceptions.swift
@@ -49,6 +49,12 @@ internal class UnableToCreateFileException: GenericException<String> {
   }
 }
 
+internal class UnableToReadHandleException: GenericException<String> {
+  override var reason: String {
+    "Unable to read from a file handle: \(param)"
+  }
+}
+
 internal class DestinationAlreadyExistsException: Exception {
   override var reason: String {
     "Destination already exists"