+++ packages/expo-file-system/ios/Next/FileSystemNextModule.swift
@@ -72,6 +72,10 @@ public final class FileSystemNextModule: Module {
         return try file.base64()
       }
 
+      Function("open") { file in
+        return try FileSystemFileHandle(file: file)
+      }
+
       Function("write") { (file, content: Either<String, TypedArray>) in
         if let content: String = content.get() {
           try file.write(content)
@@ -111,6 +115,30 @@ public final class FileSystemNextModule: Module {
 
       Property("uri") { file in
         return file.url.absoluteString
+      }
+    }
+
+    Class(FileSystemFileHandle.self) {
+      Function("readBytes") { (fileHandle, bytes: Int) in
+        try fileHandle.read(bytes)
+      }
+
+      Function("writeBytes") { (fileHandle, bytes: Data) in
+        try fileHandle.write(bytes)
+      }
+
+      Function("close") { fileHandle in
+        try fileHandle.close()
+      }
+
+      Property("offset") { fileHandle in
+        fileHandle.offset
+      }.set { (fileHandle, volume: UInt64) in
+        fileHandle.offset = volume
+      }
+
+      Property("size") { fileHandle in
+        fileHandle.size
       }
     }